concrete movement

+++ Clash/src/com/company/Models/Soldiers/Soldier.java
@@ -142,13 +142,13 @@ public abstract class Soldier {
 
     public void setX(double x) {
         this.x = x;
-        this.imageView.relocate(MapUI.mapCoordinates2PixelX((int) this.x), MapUI.mapCoordinates2PixelY((int) this.y));
+        this.imageView.relocate(MapUI.mapCoordinates2PixelX(this.x), MapUI.mapCoordinates2PixelY(this.y));
         //System.out.println(x + " " + y);
     }
 
     public void setY(double y) {
         this.y = y;
-        this.imageView.relocate(MapUI.mapCoordinates2PixelX((int) this.x), MapUI.mapCoordinates2PixelY((int) this.y));
+        this.imageView.relocate(MapUI.mapCoordinates2PixelX(this.x), MapUI.mapCoordinates2PixelY(this.y));
         //System.out.println(x + " " + y);
     }
